Add check of tag options (comma separated), add error if empty secret and marked as required
